app/requirements.txt
boto3==1.22.9
botocore==1.25.9
certifi==2021.10.8
jmespath==1.0.0
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
